requirements.txt
openai
python-dotenv
flask